requirement.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
backoff==2.2.1
branca==0.7.2
certifi==2024.7.4
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
folium==0.17.0
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
Jinja2==3.1.4
logic==0.2.3
Markdown==3.7
MarkupSafe==2.1.5
msgpack==1.1.0
multipledispatch==1.0.0
numpy==2.1.0
opencage==2.4.0
packaging==24.1
phonenumbers==8.13.43
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pygame==2.6.0
pyOpenSSL==24.2.1
redis==5.2.1
requests==2.32.3
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tk==0.1.0
toolz==0.12.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
unification==0.2.2
urllib3==2.2.2
uvicorn==0.32.1
whitenoise==6.7.0
xyzservices==2024.6.0
zope.interface==7.0.3
